requirements.txt
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1745939216646/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
diffusers==0.27.2
filelock @ file:///croot/filelock_1744281381737/work
fonttools==4.58.0
fsspec==2025.3.2
gmpy2 @ file:///croot/gmpy2_1738085463648/work
hf-xet==1.1.1
huggingface-hub==0.24.7
humanfriendly==10.0
idna @ file:///croot/idna_1714398848350/work
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2 @ file:///croot/jinja2_1741710844255/work
joblib==1.5.0
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.3
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1737039604450/work
numpy==1.26.4
onnxruntime-gpu==1.15.1
opencv-python==4.9.0.80
packaging==25.0
pandas==2.2.3
pillow @ file:///croot/pillow_1744613067434/work
pyparsing==3.2.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sympy @ file:///croot/sympy_1738108488918/work
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.44.2
triton==2.3.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1737133630106/work
vision-counter==0.1.0
zipp==3.21.0
